feat: add react scroll smooth
