+++ package.json
@@ -14,6 +14,7 @@
     "@types/node": "^16.18.82",
     "@types/react": "^18.2.56",
     "@types/react-dom": "^18.2.19",
+    "@types/react-scroll": "^1.8.10",
     "@types/three": "^0.163.0",
     "body-parser": "^1.20.2",
     "react": "^18.2.0",
