+++ src/components/menu/Menu.tsx
@@ -1,8 +1,7 @@
-import { anchorObjects } from "../../data/Data";
 import { S } from "./headerMenu/HeaderMenu.styles";
+import { anchorObjects } from "../../data/Data";
+import { Link } from "react-scroll/modules";
 import { SL } from "./Menu.styles";
-
-export const headerItems = ["Home", "About me", "Projects", "Skills", "Contacts"];
 
 export const Menu = () => {
   return (
@@ -12,13 +11,14 @@ export const Menu = () => {
           return (
             <S.HoverBox>
               <SL.ListItem key={index}>
-                <S.Link href={`#${item.href}`}>
+                <Link to={item.href} spy={true} smooth={true}>
                   <span>{item.title}</span>
-                </S.Link>
+                </Link>
               </SL.ListItem>
             </S.HoverBox>
           );
-        })};
+        })}
+        ;
       </ul>
     </S.Menu>
   );
